Continuous AP recharge

+++ Assets/Scripts/UserControl.cs
@@ -303,7 +303,7 @@ public class UserControl : MonoBehaviour
     private void showMovement(Unit u, int x, int y)
     {
         GameObject[,] objects = map.highlights;
-        bool[,] movement = AStar.movementMatrix(u.AP, map.Tiles, x, y, u.isFlying);
+		bool[,] movement = AStar.movementMatrix((int)u.AP, map.Tiles, x, y, u.isFlying);
         for(int i = 0; i < movement.GetLength(0); i++)
             for(int j = 0; j < movement.GetLength(1); j++)
             {
